Us20 (#60)
* Initialize React app

* US-20

* US-20

* US20

* US20

* created post card

* created post card

* added img modal

* added img modal

* postPage

* fixed bags

* fixed bags

* develop modalPost

* fixed bag

* fixed bags

* fixed bags

* fixed

* CreatePost

* CreatePost

* US20

* US20

* added poput to post card

* added poput to post card

* davelop post page

* develop post page

* develop post page

* develop

* fixed bugs

* fixed bugs

* create PostContent

---------
