+++ frontend/src/components/Posts/PostContent/PostContent.style.scss
